feat: :recycle: Improved separation of concerns
Moved most of the map functionality to a dedicated file.
